Se implementaron los fragments

+++ app/src/main/java/com/erwiin21mp/cinemovilplus/domain/model/ContentModel.kt
@@ -15,7 +15,7 @@ data class ContentModel(
     val contentUrl: String,
     val rating: Int,
     val platformsList: List<String>,
-    val uploadDate: Int,
+    val uploadDate: Long,
     val producersList: List<String>,
     val type: String,
     val distribution: String,
